Make use of the caffe ``snapshot`` and ``copy_from`` methods if exposed to Python.
With this update, caffe's internal ``snapshot`` and ``copy_from``
methods are used instead of the ``save`` method of the Net and
barrista's ``load_blobs_from`` added method. This is a step towards out-
of-the-box compatibility with caffe. The last remaining required method
in C++ code is the ``InitForNet`` method for the solver.

The save paths for the Checkpointer monitor change, since the path
layout is now similar to caffe. '_iter_' is added to the saved files,
and they count now processed batches instead of plain iterations.