Enhanced the CyclingDataMonitor, added ResizingMonitor and RotatingMirroringMonitor.
This update comes with plenty of new features. The CyclingDataMonitor
has been enhanced to handle lists as well and accept preprocessing flags
for resizing or padding (this feature can now used directly for the
Net.fit and solver.fit method as well).

The ResizingMonitor offers many options to scale data, also randomly for
data enhancement, and to automatically adjust a fully convolutional
network's input size to the samples.

The RotatingMirroringMonitor offers additional data augmentation
features for rotation and mirroring.